Refactor: Move OutdatednessRules into its own file
